[track] Remove usage of `namespace`.
Babel 7’s TS plugin does not support it, but out linter also doesn’t
so good riddance altogether.
